consolidate on hashable float wrapper
