+++ src/daft-core/src/utils/mod.rs
@@ -1,7 +1,6 @@
 pub mod arrow;
 pub mod display_table;
 pub mod hashable_float_wrapper;
-pub mod orderable_float_wrapper;
 pub mod supertype;
 
 #[macro_export]
