--- src/daft-core/src/utils/orderable_float_wrapper.rs
@@ -1,43 +1,0 @@
-use std::cmp::Ordering;
-
-pub struct FloatWrapper<T>(pub T);
-
-impl<T: PartialEq> PartialEq for FloatWrapper<T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.eq(&other.0)
-    }
-}
-
-impl Eq for FloatWrapper<f32> {}
-
-impl Eq for FloatWrapper<f64> {}
-
-// This implementation of Ord considers NaNs to be equal to each other, and less than any other value.
-
-impl<T: PartialOrd> PartialOrd for FloatWrapper<T> {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        self.0.partial_cmp(&other.0)
-    }
-}
-
-impl Ord for FloatWrapper<f32> {
-    fn cmp(&self, other: &Self) -> Ordering {
-        match (self.0.is_nan(), other.0.is_nan()) {
-            (true, true) => Ordering::Equal,
-            (true, false) => Ordering::Less,
-            (false, true) => Ordering::Greater,
-            (false, false) => self.0.partial_cmp(&other.0).unwrap(),
-        }
-    }
-}
-
-impl Ord for FloatWrapper<f64> {
-    fn cmp(&self, other: &Self) -> Ordering {
-        match (self.0.is_nan(), other.0.is_nan()) {
-            (true, true) => Ordering::Equal,
-            (true, false) => Ordering::Less,
-            (false, true) => Ordering::Greater,
-            (false, false) => self.0.partial_cmp(&other.0).unwrap(),
-        }
-    }
-}
